Struct neko::dynamic::library::Library

pub struct Library { /* fields omitted */ }

The struct Library is a table of callback.

Methods

impl Library

fn new(path: PathBuf, index: i64, state: &LibraryState) -> Result<Self>

The constructor method new returns a interface for a extern library.

fn as_path_buf(&self) -> &PathBuf

The accessor method as_path_buf return address of library.

fn get_priority(&self) -> i64

The accessor method get_priority return level's priority of library.

fn is_unmounted(&self) -> bool

fn install(&self, state: &LibraryState)

The method install call the extern function if defined when the library is installed.

fn uninstall(&self, state: &LibraryState)

The method uninstall call the extern function if defined when the library is uninstalled.

fn start(&self, state: &LibraryState)

The method start call the extern function if defined when the library is mounted.

fn end(&self, state: &LibraryState)

The method end call the extern function if defined when the library is unmounted.

fn idle(&self, state: &LibraryState)

The method end call the extern function if defined.

fn process(&self, state: &LibraryState, taskname: &[c_uchar], pid: c_int)

The method process call the extern function if defined when the child current process as been canged.

fn command(&self, state: &LibraryState, line: &[c_uchar])

The method command call the extern function if defined when a command line is outputed to the terminal.

fn key_unicode_down(&self, state: &LibraryState, code: c_ulonglong)

The method key_unicode_down call the extern function if defined when a key is pressed.

fn key_string_down(&self, state: &LibraryState, text: &[c_uchar])

The method key_string_down call the extern function if defined when a text is pasted or pressed.

fn key_repeat_down(&self, state: &LibraryState, repeat: c_ulong)

The method key_repeat_down call the extern function if defined when a key is held. It determines the time before it begins to output the held character within a given interval.

fn key_interval_down(&self, state: &LibraryState, interval: c_longlong)

The method key_interval_down call the extern function if defined when a key is held. It determines the time between two outputs repetitions of the held character. (Triggered by key_repeat_down())

fn mouse_pressed(&self, state: &LibraryState, code: c_uint, xy: [c_ushort; 2])

The method mouse_pressed call the extern function if defined when the mouse is pressed.

fn mouse_released(&self, state: &LibraryState, code: c_uint, xy: [c_ushort; 2])

The method mouse_released call the extern function if defined when the mouse is released.

fn input(&self, state: &LibraryState, text: &[c_uchar])

The method input call the extern function if defined when something is inputted to the terminal, whatever it is.

fn output(&self, state: &LibraryState, text: &[c_uchar])

The method output call the extern function if defined. when something is outputted to the terminal, whatever it is.

fn resized(&self, state: &LibraryState, size: &Winszed)

The method resized call the extern function if defined when the window is resized.

fn call(&self, state: &LibraryState, event: &ShellState)

The method call will read the ShellState to call an adapted extern function if defined.