Struct neko::dynamic::Compositer

pub struct Compositer { /* fields omitted */ }

The struct Compositer is a heap of a double tuple of a dynamic libraries and a priority order.

Methods

impl Compositer

fn new() -> Result<Self>

The constructor new returns a Compositer prepared with the library root.

fn get_git(&self) -> Result<PathBuf>

The accessor method get_git returns the git sub-directory.

fn get_lib(&self) -> Result<PathBuf>

The accessor method get_lib returns the lib sub-directory.

fn git_with_lib(&mut self) -> Result<(PathBuf, PathBuf)>

The accessor method git_with_lib returns a couple of git and lib sub-repositories.

fn get_manifest(&self, repository: &PathBuf) -> Result<Table>

The accessor method get_manifest returns a dictionary. @repository: $HOME/.neko/git/arukana@libnya.

fn mount<S: AsRef<OsStr>>(
&mut self,
libraryname: &S,
priority: Option<i64>
) -> Result<()>

The method mount adds a new library to the heap's compositer. @ libraryname: arukana@libnya. @ priority: Some(-1) or None for zero by default.

fn unmount<S: AsRef<OsStr>>(&mut self, libraryname: S) -> Result<()>

The method unmount removes library from the queue. @ libraryname: arukana@libnya.

fn build<S: AsRef<OsStr> + AsRef<Path>>(
&mut self,
source: &PathBuf,
sub: S
) -> Result<()>

The method build makes and adds a dynamic library to SPEC_MANIFEST's destination. @ source: $HOME/.neko/git/Arukana@libnya. @ sub: arukana@libnya.

fn dependency(&mut self, source: &PathBuf) -> Result<()>

The method dependency lists the dependencies from repository dynamic library and install. @ source: $HOME/.neko/git/Arukana@libnya.

fn install(&mut self, repo: &str) -> Result<()>

The methodd install clones and makes a dynamic library from repository and recursive call the dependencies. @ repo: https://github.com/Arukana/libnya.git.

fn update(&mut self, libraryname: &str) -> Result<()>

The method update hard-resets the master branch to last commit. @ libraryname: arukana@libnya.

fn uninstall<S: AsRef<OsStr>>(&mut self, libraryname: &S) -> Result<()>

The method uninstall removes library from the filesystem with the source. @libraryname: arukana@libnya.

fn get_state(&self) -> &LibraryState

fn set_tooltip_message(&mut self, text: String)

fn set_tooltip_cardinal(&mut self, sheet: Relative)

fn set_persona_sheet(&mut self, sheet: Sheet)

fn set_persona_position(&mut self, position: Position)

fn resized(&mut self, size: &Winszed)

The general method call according to the state will run the evenement functions by library group.

fn call(&mut self, event: &ShellState)

The general method call according to the state will run the evenement functions by library group.