Struct neko::graphic::Sprite

pub struct Sprite { /* fields omitted */ }

Methods

impl Sprite

fn explicite_emotion(&mut self, change: &[[Tuple; 50]; 16])

fn insert_list(&mut self, duration: i64, posture: &Sheet, source: &[Tuple])

The function insert_list push a new draw from a list of tuple of emotion by part.

fn extend(&mut self, texels: &HashMap<Tuple, Vec<Texel>, RandomState>)

The function extend extends the local dictionary of texel.

fn current(&self) -> Option<(&Emotion, &Texel)>

fn set_current(&mut self, cell: (&Emotion, &Vec<Texel>)) -> Option<()>

fn get_posture(&self) -> Option<&Sheet>

fn get_current_draw(&self) -> Option<&Draw>

fn add_position(&mut self, position: usize) -> Option<()>

The mutator method add_position_draw increments the position of the draw sheet cursor.

fn sub_position(&mut self, position: usize) -> Option<()>

The mutator method sub_position decrements the position of the draw sheet cursor.

fn add_position_draw(&mut self, position: usize) -> Option<()>

The mutator method add_position_draw increments the position of the cell board cursor.

fn sub_position_draw(&mut self, position: usize) -> Option<()>

The mutator method sub_position_draw decrements the position of the cell board cursor.