Enum neko::prelude::Code

#[repr(u32)]

pub enum Code {
    Left,
    Right,
    Wheel,
    WheelUp,
    WheelDown,
    LeftDrag,
    WheelDrag,
    RightDrag,
    ShiftLeft,
    ShiftWheel,
    ShiftRight,
    ShiftLeftDrag,
    ShiftWheelDrag,
    ShiftRightDrag,
    CtrlLeft,
    CtrlWheel,
    CtrlRight,
    CtrlWheelUp,
    CtrlWheelDown,
    CtrlLeftDrag,
    CtrlWheelDrag,
    CtrlRightDrag,
    ShiftCtrlLeft,
    ShiftCtrlWheel,
    ShiftCtrlRight,
    ShiftCtrlLeftDrag,
    ShiftCtrlWheelDrag,
    ShiftCtrlRightDrag,
    CmdLeft,
    CmdWheel,
    CmdRight,
    CmdWheelUp,
    CmdWheelDown,
    CmdLeftDrag,
    CmdWheelDrag,
    CmdRightDrag,
    CmdShiftLeft,
    CmdShiftWheel,
    CmdShiftRight,
    CmdShiftLeftDrag,
    CmdShiftWheelDrag,
    CmdShiftRightDrag,
    CmdCtrlLeft,
    CmdCtrlWheel,
    CmdCtrlRight,
    CmdCtrlWheelUp,
    CmdCtrlWheelDown,
    CmdCtrlLeftDrag,
    CmdCtrlWheelDrag,
    CmdCtrlRightDrag,
    CmdShiftCtrlLeft,
    CmdShiftCtrlWheel,
    CmdShiftCtrlRight,
    CmdShiftCtrlLeftDrag,
    CmdShiftCtrlWheelDrag,
    CmdShiftCtrlRightDrag,
}

Variants
Left

The left mouse button is pressed.

Right

The right mouse button is pressed.

Wheel

The mouse wheel button is pressed.

WheelUp

The mouse wheel is going up.

WheelDown

The mouse wheel is going down.

LeftDrag

The left mouse button is held while moving pointer.

WheelDrag

The wheel mouse button is held while moving pointer.

RightDrag

The right mouse button is held while moving pointer.

ShiftLeft

The left mouse button is pressed while helding Shift.

ShiftWheel

The wheel mouse button is pressed while helding Shift.

ShiftRight

The right mouse button is pressed while helding Shift.

ShiftLeftDrag

The left mouse button and Shift are held while moving pointer.

ShiftWheelDrag

The wheel mouse button and Shift are held while moving pointer.

ShiftRightDrag

The right mouse button and Shift are held while moving pointer.

CtrlLeft

The left mouse button is pressed while helding Ctrl

CtrlWheel

The wheel mouse button is pressed while helding Ctrl

CtrlRight

The right mouse button is pressed while helding Ctrl

CtrlWheelUp

The mouse wheel is going up while helding Ctrl

CtrlWheelDown

The mouse wheel is going down while helding Ctrl

CtrlLeftDrag

The left mouse button and Ctrl are held while moving pointer

CtrlWheelDrag

The wheel mouse button and Ctrl are held while moving pointer

CtrlRightDrag

The right mouse button and Ctrl are held while moving pointer

ShiftCtrlLeft

The left mouse button is pressed while Ctrl and Shift are held

ShiftCtrlWheel

The wheel mouse button is pressed while Ctrl and Shift are held

ShiftCtrlRight

The right mouse button is pressed while Ctrl and Shift are held

ShiftCtrlLeftDrag

The left mouse button, Ctrl and Shift are held while moving pointer

ShiftCtrlWheelDrag

The wheel mouse button, Ctrl and Shift are held while moving pointer

ShiftCtrlRightDrag

The right mouse button, Ctrl and Shift are held while moving pointer

CmdLeft

The left mouse button is pressed while helding Command

CmdWheel

The wheel mouse button is pressed while helding Command

CmdRight

The right mouse button is pressed while helding Command

CmdWheelUp

The mouse wheel is going up while helding Command

CmdWheelDown

The mouse wheel is going down while helding Command

CmdLeftDrag

The left mouse button and Command are held while moving pointer

CmdWheelDrag

The wheel mouse button and Command are held while moving pointer

CmdRightDrag

The right mouse button and Command are held while moving pointer

CmdShiftLeft

The left mouse button is pressed while helding Command and Shift

CmdShiftWheel

The wheel mouse button is pressed while helding Command and Shift

CmdShiftRight

The right mouse button is pressed while helding Command and Shift

CmdShiftLeftDrag

The left mouse button, Shift and Command are held while moving pointer

CmdShiftWheelDrag

The wheel mouse button, Shift and Command are held while moving pointer

CmdShiftRightDrag

The right mouse button, Shift and Command are held while moving pointer

CmdCtrlLeft

The left mouse button is pressed while helding Command and Ctrl

CmdCtrlWheel

The wheel mouse button is pressed while helding Command and Ctrl

CmdCtrlRight

The right mouse button is pressed while helding Command and Ctrl

CmdCtrlWheelUp

The mouse wheel is going up while helding Command and Ctrl

CmdCtrlWheelDown

The mouse wheel is going down while helding Command and Ctrl

CmdCtrlLeftDrag

The left mouse button, Ctrl and Command are held while moving pointer

CmdCtrlWheelDrag

The wheel mouse button, Ctrl and Command are held while moving pointer

CmdCtrlRightDrag

The right mouse button, Ctrl and Command are held while moving pointer

CmdShiftCtrlLeft

The left mouse button is pressed while helding Command, Shift and Ctrl

CmdShiftCtrlWheel

The wheel mouse button is pressed while helding Command, Shift and Ctrl

CmdShiftCtrlRight

The right mouse button is pressed while helding Command, Shift and Ctrl

CmdShiftCtrlLeftDrag

The left mouse button, Ctrl, Shift and Command are held while moving pointer

CmdShiftCtrlWheelDrag

The wheel mouse button, Ctrl, Shift and Command are held while moving pointer

CmdShiftCtrlRightDrag

The right mouse button, Ctrl, Shift and Command are held while moving pointer

Methods

impl Code

fn new(action: u8) -> Result<Code, CodeError>