Enum neko::pty::Key

pub enum Key {
    Char(u64),
    Str(In),
}

Variants
Char(u64)

Unicode characters.

Str(In)

Unicode strings.

Methods

impl Key

fn from_utf8(buf: [u8; 4]) -> Key

The constructor method from_utf8 returns a UTF-8 parsed Key.

fn is_utf8(&self) -> Option<char>

fn is_up(&self) -> bool

The accessor method is_up returns an Option for the Up Key.

fn is_down(&self) -> bool

The accessor method is_down returns an Option for the down Key.

fn is_right(&self) -> bool

The accessor method is_right returns an Option for the right Key.

fn is_left(&self) -> bool

The accessor method is_left returns an Option for the left Key.

fn is_delete(&self) -> bool

The accessor method is_delete returns an Option for the delete Key.

fn is_backspace(&self) -> bool

The accessor method is_backspace returns an Option for the backspace Key.

fn is_escape(&self) -> bool

The accessor method is_escape returns an Option for the delete Key.

fn is_end(&self) -> bool

The accessor method is_end returns an Option for the delete Key.

fn is_home(&self) -> bool

The accessor method is_home returns an Option for the delete Key.

fn is_pageup(&self) -> bool

The accessor method is_pageup returns an Option for the Page Up Key.

fn is_pagedown(&self) -> bool

The accessor method is_pagedown returns an Option for the Page Down Key.

fn is_c0(&self) -> bool

fn is_null(&self) -> bool

fn is_start_heading(&self) -> bool

fn is_start_text(&self) -> bool

fn is_end_text(&self) -> bool

fn is_end_transmission(&self) -> bool

fn is_enquiry(&self) -> bool

fn is_acknowledge(&self) -> bool

fn is_bell(&self) -> bool

fn is_horizontal_tabulation(&self) -> bool

fn is_enter(&self) -> bool

The accessor method is_enter returns an Option for the Enter Key.

fn is_line_feed(&self) -> bool

fn is_vertical_tabulation(&self) -> bool

fn is_form_feed(&self) -> bool

fn is_carriage_return(&self) -> bool

fn is_shift_out(&self) -> bool

fn is_shift_in(&self) -> bool

fn is_data_link_escape(&self) -> bool

fn is_device_control_one(&self) -> bool

fn is_device_control_two(&self) -> bool

fn is_device_control_three(&self) -> bool

fn is_device_control_four(&self) -> bool

fn is_negative_acknowledge(&self) -> bool

fn is_synchronous_idle(&self) -> bool

fn is_end_transmission_block(&self) -> bool

fn is_cancel(&self) -> bool

fn is_end_of_medium(&self) -> bool

fn is_substitute(&self) -> bool

fn is_file_separator(&self) -> bool

fn is_group_separator(&self) -> bool

fn is_record_separator(&self) -> bool

fn is_unit_separator(&self) -> bool

fn is_space(&self) -> bool

fn is_c1(&self) -> bool

fn as_input(&self) -> (In, usize)