Struct neko::dynamic::Compositer

pub struct Compositer(_);

The struct Compositer is a heap of a double tuple of a dynamic libraries and a priority order.

Methods

impl Compositer

fn source(&self) -> PathBuf

The accessor method source returns the source directory where found the git and lib sub-directories.

fn mount<S: AsRef<OsStr>>(&mut self, libraryname: S, priority: Option<usize>) -> Result<()>

The method mount adds a new library to the heap's compositer.

fn mount_from_dir(&mut self, directory: &str, priority: Option<usize>) -> Result<()>

The method mount_from_dir adds a new library to the heap's compositer from a directory.

fn mount_from_git(&mut self, repository: &str, priority: Option<usize>) -> Result<()>

The method mount_from_git adds a new library to the heap's compositer from a git repository.

fn start(&self)